Kbuild: Select LD_DEAD_CODE_DATA_ELIMINATION with LTO_CLANG
This selection being the default behavior for Clang LTO has been
reverted upstream over concerns that the use of -gc-sections
carries a greater potential risk of breakage.

That being said, as someone who has been using these features in
tandem for several years to no ill effect, this is a risk that I
am willing to take in order to trim the fat from LTO's thick
kernel images and potentially reduce boot times.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>